requirements.txt
protobuf==4.23.2
python-memcached==1.59